Fix reruns

+++ scripts/grow_root.sh
@@ -2,6 +2,11 @@
 
 set -e
 
+if [ -f /local/complete_grow ]; then
+    echo "Already completed, skipping"
+    exit 0
+fi
+
 apt-get update
 apt-get install -y cloud-guest-utils
 
@@ -10,3 +15,5 @@ ROOT_DISK=/dev/"$(lsblk -no PKNAME "$(findmnt -no SOURCE /)")"
 
 growpart "$ROOT_DISK" "$ROOT_PART"
 resize2fs "$(findmnt / -no SOURCE)"
+
+touch /local/complete_grow
